[Rename] SignInput 위치 변경 - level-2 디렉토리로 이동
